Add methods to visit class

+++ cos/lab05/lab05/Models/Visit.cs
@@ -5,9 +5,9 @@ public class Visit
     public String date { get; set; }
     public Animal animal { get; set; }
     public String brief { get; set; }
-    public int cost { get; set; }
+    public double cost { get; set; }
 
-    public Visit(string date, Animal animal, string brief, int cost)
+    public Visit(string date, Animal animal, string brief, double cost)
     {
         this.date = date;
         this.animal = animal;
